Elide v0.0.1 Elide.Auth

Functions

authenticate_user(conn, opts)

Returns the user if exists in conn, otherwise redirect user to home

authorize_url!(arg1)

Authorizes received token through google api

For now we only support google as oauth provider

call(conn, repo)

Gets the logged in user from session

get_user_details!(binary, code)

Gets user’s details based on google oauth response

init(opts)

login(conn, user)

Prepares client’s session to keep the logged in user

logout(conn)

Cleans up client’s session