Elide v0.0.1 Elide.Domain

Functions

by_address(domain_name)

changeset(model, params \\ :empty)

Creates a changeset based on the model and params.

If no params are provided, an invalid changeset is returned with no validation performed.

default_domain()