Elide v0.0.1 Elide.Elink

Functions

by_slug(slug)

Returns query based on short hash url

This query should return only one Elink item

changeset(model, params \\ :empty)

Creates a changeset based on the model and params.

If no params are provided, an invalid changeset is returned with no validation performed.

get_details_by_slug(slug)

salt()

short_url(elink)

Returns full short link

slug(elink)

Returns short url hash based on elink