Elide v0.0.1 Elide.ElinkSeqServer

Provides elink sequence for given domain

The idea is to keep the elinks as short as possible by creating elink hash per domain. This module is responsible to provide unique hash per domain

> It’s not safe to use this module in distributed environment

Functions

get_sequence(domain)

Gets the last used sequence for given domain

init()

Initialize ElinkSeqServer state

nextval(domain_id)

Returns the next sequence id for given domain_id

iex> Elide.ElinkSeqServer.nextval(20)
1
iex> Elide.ElinkSeqServer.nextval(20)
2

Returns the next sequence id for given domain

iex> domain = %Elide.Domain{id: 12}
iex> Elide.ElinkSeqServer.nextval(domain)
1
iex> Elide.ElinkSeqServer.nextval(domain)
2