Elide v0.0.1 Elide.RateLimiter

Keep track of api usage per given key, like ip address or user id

iex> {:ok, cache} = RateLimiter.start_link(
...>   [api_rate_limit: 2, ttl: :timer.hours(1), ttl_check: :timer.minutes(1)])
iex> RateLimiter.check_limit!("127.0.0.1", cache)
true
iex> RateLimiter.check_limit!("127.0.0.1", cache)
true
iex> RateLimiter.check_limit!("127.0.0.1", cache)
false

Functions

check_limit!(limitation_key)

check_limit!(limitation_key, pid)

Checkes if the given key has reached it’s limit on given period of time

start_link()

start_link(options, gen_server_options \\ [])

Creates a cache process, with capability of ttl for each cache item

See ConCache for more details.