Elide v0.0.1 Elide.Router.Helpers

Module with named helpers generated from Elide.Router.

Functions

elink_api_path(conn_or_endpoint, action)

elink_api_path(conn_or_endpoint, action, params)

elink_api_url(conn_or_endpoint, action)

elink_api_url(conn_or_endpoint, action, params)

elink_path(conn_or_endpoint, action)

elink_path(conn_or_endpoint, action, slug)

elink_path(conn_or_endpoint, action, slug, params)

elink_url(conn_or_endpoint, action)

elink_url(conn_or_endpoint, action, slug)

elink_url(conn_or_endpoint, action, slug, params)

organization_membership_path(conn_or_endpoint, action, organization_id)

organization_membership_path(conn_or_endpoint, action, organization_id, params)

organization_membership_path(conn_or_endpoint, action, organization_id, id, params)

organization_membership_url(conn_or_endpoint, action, organization_id)

organization_membership_url(conn_or_endpoint, action, organization_id, params)

organization_membership_url(conn_or_endpoint, action, organization_id, id, params)

organization_path(conn_or_endpoint, action)

organization_path(conn_or_endpoint, action, params)

organization_path(conn_or_endpoint, action, id, params)

organization_url(conn_or_endpoint, action)

organization_url(conn_or_endpoint, action, params)

organization_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

token_path(conn_or_endpoint, action)

token_path(conn_or_endpoint, action, params)

token_path(conn_or_endpoint, action, id, params)

token_url(conn_or_endpoint, action)

token_url(conn_or_endpoint, action, params)

token_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action, provider)

user_path(conn_or_endpoint, action, provider, params)

user_url(conn_or_endpoint, action, provider)

user_url(conn_or_endpoint, action, provider, params)